// src/mail/type/mail.enum.ts
export enum EnumMailStatus {
  Draft = 'draft',
  Pending = 'pending',
  Processing = 'processing',
  Success = 'success',
  Error = 'error',
  Warning = 'warning',
}

export enum EnumMailType {
  Question = 'question',
  EmailOtp = 'email-otp',
  EmailLinkOtp = 'email-link-otp',
  PhoneOtp = 'phone-otp',
  Information = 'information',
  Advert = 'advert',
}